+++ packages/core/CHANGELOG.md
@@ -1,1 +1,12 @@
 # @rspress/core
+
+## 0.0.1
+
+### Patch Changes
+
+- db2d3cd: feat: add plugins
+- Updated dependencies [db2d3cd]
+  - @rspress/plugin-auto-nav-sidebar@0.0.1
+  - @rspress/plugin-last-updated@0.0.1
+  - @rspress/plugin-medium-zoom@0.0.1
+  - @rspress/shared@0.0.1