+++ packages/plugin-preview/CHANGELOG.md
@@ -1,1 +1,9 @@
 # @rspress/plugin-preview
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [db2d3cd]
+  - @rspress/core@0.0.1
+  - @rspress/shared@0.0.1