+++ packages/plugin-typedoc/CHANGELOG.md
@@ -1,1 +1,10 @@
 # @rspress/plugin-typedoc
+
+## 0.0.1
+
+### Patch Changes
+
+- db2d3cd: feat: add plugins
+- Updated dependencies [db2d3cd]
+- Updated dependencies [39a0907]
+  - rspress@0.0.1